Move rent debit out of bank
