+++ runtime/src/accounts.rs
@@ -13,12 +13,10 @@ use {
         },
         accounts_update_notifier_interface::AccountsUpdateNotifier,
         ancestors::Ancestors,
-        bank::{
-            Bank, NonceFull, NonceInfo, RentDebits, TransactionCheckResult,
-            TransactionExecutionResult,
-        },
+        bank::{Bank, NonceFull, NonceInfo, TransactionCheckResult, TransactionExecutionResult},
         blockhash_queue::BlockhashQueue,
         rent_collector::RentCollector,
+        rent_debit::RentDebits,
         storable_accounts::StorableAccounts,
         system_instruction_processor::{get_system_account_kind, SystemAccountKind},
         transaction_error_metrics::TransactionErrorMetrics,
